feat: Add ConferenceTable component
A new component, ConferenceTable, has been added to the pieces directory. This component will be used to display conference tables in the application. It has been imported and used in the Schedule page to show the conference schedule. This commit adds the necessary files and updates the index.jsx file in the pieces directory.